app/requirements.txt
# IDS Kafka Pipeline - Dependencies

# Core
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# ML
scikit-learn>=1.3.0
joblib>=1.3.0

# Kafka (optionnel - necessaire pour mode producer/consumer)
kafka-python>=2.0.2
